+++ NEWS.md
@@ -8,6 +8,7 @@
 ## Changes to functions
 
 * `tab_model()` gets a `strings`-argument, as convenient shortcut for setting column-header strings.
+* `tab_model()` and `plot_model()` get additional arguments `vcov.fun`, `vcov.type` and `vcov.args` that are passed down to `sjstats::robust()`, to calculate different types of (clustered) robust standard errors.
 
 ## Bug fixes